refactor(geSourceImageUrl): unify temp fix
This fix is currently used in two places,
so unifying it should help with future removal.
