+++ dotcom-rendering/src/web/components/Carousel.importable.tsx
@@ -17,6 +17,7 @@ import type { Palette } from '../../types/palette';
 import type { TrailType } from '../../types/trails';
 import { decidePalette } from '../lib/decidePalette';
 import { formatAttrString } from '../lib/formatAttrString';
+import { getSourceImageUrl } from '../lib/getSourceImageUrl_temp_fix';
 import { getZIndex } from '../lib/getZIndex';
 import { Card } from './Card/Card';
 import { LI } from './Card/components/LI';
@@ -429,36 +430,6 @@ const HeaderAndNav = ({
 	</div>
 );
 
-/**
- * Find the master URL for an image if has been generated by frontend.
- *
- * Temporary fix while `/most-read/*.json` do not return the master URLS.
- * This is undoing the work that happens in `Picture.tsx`’s `generateImageURL`,
- * so it should be removed in favour of a the upstream data returning the
- * source of the master image, rather than a specific signed crop
- */
-const getImageUrl = (source: string | undefined) => {
-	if (!source) return undefined;
-	const url = new URL(source);
-
-	const buckets = ['/img/media/', '/img/static/', '/img/uploads/'] as const;
-	const prefix = buckets.find((bucket) => url.pathname.startsWith(bucket));
-
-	if (url.hostname === 'i.guim.co.uk' && prefix) {
-		const path = url.pathname.replace(prefix, '');
-		switch (prefix) {
-			case '/img/media/':
-				return `https://media.guim.co.uk/${path}`;
-			case '/img/static/':
-				return `https://static.guim.co.uk/${path}`;
-			case '/img/uploads/':
-				return `https://uploads.guim.co.uk/${path}`;
-		}
-	}
-
-	return source;
-};
-
 export const Carousel = ({ heading, trails, onwardsSource, format }: Props) => {
 	const palette = decidePalette(format);
 	const carouselRef = useRef<HTMLUListElement>(null);
@@ -675,7 +646,7 @@ export const Carousel = ({ heading, trails, onwardsSource, format }: Props) => {
 						const discussionId =
 							shortUrl && new URL(shortUrl).pathname;
 
-						const imageUrl = getImageUrl(image);
+						const imageUrl = image && getSourceImageUrl(image);
 
 						return (
 							<CarouselCard
